+++ folder-paths.js
@@ -1,0 +1,13 @@
+const path = require("path");
+
+function defineFolderPaths()
+{
+	var defineRes = {};	
+	defineRes["input"] = path.join(".", "input-data");
+	defineRes["output"] = path.join(".", "output-sql");
+	
+	return defineRes;
+}
+
+
+module.exports = defineFolderPaths();