+++ changelog.md
@@ -1,4 +1,7 @@
 # Changelog
 
-**./output-sql/**
-* New folder - Generated SQL files will be saved here.
+**./folder-paths.js**
+* New file
+	* Contains paths to input and output folders.
+	* This does not define the full names of files.
+	* They will be generated separately as needed.